refactor ds2, cli, server
