+++ paddlespeech/s2t/exps/deepspeech2/bin/export.py
@@ -32,11 +32,9 @@ def main(config, args):
 
 if __name__ == "__main__":
     parser = default_argument_parser()
-    # save jit model to 
+    # save jit model to
     parser.add_argument(
         "--export_path", type=str, help="path of the jit model to save")
-    parser.add_argument(
-        "--model_type", type=str, default='offline', help="offline/online")
     parser.add_argument(
         '--nxpu',
         type=int,
@@ -44,7 +42,6 @@ if __name__ == "__main__":
         choices=[0, 1],
         help="if nxpu == 0 and ngpu == 0, use cpu.")
     args = parser.parse_args()
-    print("model_type:{}".format(args.model_type))
     print_arguments(args)
 
     # https://yaml.org/type/float.html
