+++ paddlespeech/s2t/exps/deepspeech2/bin/test_export.py
@@ -39,8 +39,6 @@ if __name__ == "__main__":
     parser.add_argument(
         "--export_path", type=str, help="path of the jit model to save")
     parser.add_argument(
-        "--model_type", type=str, default='offline', help='offline/online')
-    parser.add_argument(
         '--nxpu',
         type=int,
         default=0,
@@ -50,7 +48,6 @@ if __name__ == "__main__":
         "--enable-auto-log", action="store_true", help="use auto log")
     args = parser.parse_args()
     print_arguments(args, globals())
-    print("model_type:{}".format(args.model_type))
 
     # https://yaml.org/type/float.html
     config = CfgNode(new_allowed=True)
