+++ paddlespeech/s2t/exps/deepspeech2/bin/train.py
@@ -32,15 +32,12 @@ def main(config, args):
 if __name__ == "__main__":
     parser = default_argument_parser()
     parser.add_argument(
-        "--model_type", type=str, default='offline', help='offline/online')
-    parser.add_argument(
         '--nxpu',
         type=int,
         default=0,
         choices=[0, 1],
         help="if nxpu == 0 and ngpu == 0, use cpu.")
     args = parser.parse_args()
-    print("model_type:{}".format(args.model_type))
     print_arguments(args, globals())
 
     # https://yaml.org/type/float.html
